Refactored assignment submission form
The assignment submission form is now it's own page and can be included
in the different pages. This makes it easier to modify the assignment
submission process (images/applet/etc.) look and feel without constant
switching between pages.

The refactor also made it easier to implement jar uploads for Professor
assignments, which addresses issue
https://github.com/kelvinhsung/GTCS12/issues/29 and
https://github.com/kelvinhsung/GTCS12/issues/30.