Refactored course-selector into separate file
The course-selector code was shared among multiple pages and didnt need
to be duplicated. Separating it into a separate file allows it to be
modularly added to pages. The pages still need to handle the
initialization of the selectors state however.

TODO: fix the isOwner bug by adding functionality to check if the course
belongs to the professor